+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * Add 'About Us' page ([#591](https://gitlab.com/yale-sdmp/vesta/issues/591)).
 * Update feature specs to use navigation ([#399] (https://gitlab.com/yale-sdmp/vesta/issues/399)).
 * Add a suite decorator ([#581] (https://gitlab.com/yale-sdmp/vesta/issues/581)).
+* Add ability to reassign suite in the suite selection or results phase ([#969] (https://gitlab.com/yale-sdmp/vesta/issues/969)).
 
 ### Changed
 * Sort groups in results and suite selection tables by lottery number ([#968](https://gitlab.com/yale-sdmp/vesta/issues/968)).